+++ packages/base/src/components/MinusButton/MinusButton.stories.tsx
@@ -1,0 +1,10 @@
+import type { StoryFn } from '@storybook/react';
+import MinusButton from './MinusButton';
+import type { MinusBoxProps } from './MinusButton.types';
+
+export default {
+  title: 'COMPONENTS/MinusButton',
+  component: MinusButton,
+};
+
+export const Playground: StoryFn<MinusBoxProps> = (args) => <MinusButton {...args} />;