+++ packages/base/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @shoplflow/base
 
+## 0.9.0
+
+### Minor Changes
+
+- [#92](https://github.com/shopl/shoplflow/pull/92) [`4c2d1ce`](https://github.com/shopl/shoplflow/commit/4c2d1ceb90b5cb248a954ef6be4f895cf34e86f8) Thanks [@dev-bomdong](https://github.com/dev-bomdong)! - Modal,Backdrop,Text 수정 및 Switch 컴포넌트명 변경
+
+## 0.7.0
+
+### Minor Changes
+
+- [#69](https://github.com/shopl/shoplflow/pull/69) [`577adc9`](https://github.com/shopl/shoplflow/commit/577adc98a316bede622cdc6135c81ac0974056d5) Thanks [@dev-bomdong](https://github.com/dev-bomdong)! - add Minus button and Toggle
+
+- [#90](https://github.com/shopl/shoplflow/pull/90) [`961952b`](https://github.com/shopl/shoplflow/commit/961952b76f90d9112d246307a7502af92ead53ff) Thanks [@dev-bomdong](https://github.com/dev-bomdong)! - Modal, Backdrop style 수정 및 Switch 컴포넌트명 변경
+
 ## 0.6.0
 
 ### Minor Changes